Move version parsing tests into misc.
